feat: Add conditional check for NIC type (FrontendNic) in DeleteState method to enhance endpoint deletion logic

+++ network/manager.go
@@ -772,11 +772,13 @@ func (nm *networkManager) DeleteState(epInfos []*EndpointInfo) error {
 
 	if nm.IsStatelessCNIMode() {
 		for _, epInfo := range epInfos {
-			response, err := nm.CnsClient.DeleteEndpointState(context.TODO(), epInfo.ContainerID)
-			if err != nil {
-				return errors.Wrapf(err, "Delete endpoint API returned with error for endpoint %s", epInfo.ContainerID)
+			if epInfo.NICType == cns.NodeNetworkInterfaceFrontendNIC {
+				response, err := nm.CnsClient.DeleteEndpointState(context.TODO(), epInfo.ContainerID)
+				if err != nil {
+					return errors.Wrapf(err, "Delete endpoint API returned with error for endpoint %s", epInfo.ContainerID)
+				}
+				logger.Info("Delete endpoint API returned", zap.String("endpointID", epInfo.ContainerID), zap.String("returnCode", response.ReturnCode.String()))
 			}
-			logger.Info("Delete endpoint API returned", zap.String("endpointID", epInfo.ContainerID), zap.String("returnCode", response.ReturnCode.String()))
 		}
 		return nil
 	}
